app_dev/NavSwap_AI_Service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
joblib>=1.3.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart
google-generativeai>=0.3.0
